requirements.txt
asgiref==3.7.2
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
Faker==20.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
